allow reuse of CKAssets
This adds a CKRereferenceAssetOperation, which returns
a CKAsset that can be attached to a new CKRecord, thus allowing reuse
of CKAssets.

While OpenCloudKit does still not support uploading new assets,
this allows it to reuse existing ones by calling iCloud's rereference
action. It returns enough information to put that CKAsset on a new record.

Maybe this could be built into `normal` CKRecord's save operations,
but for now, this solves our needs. :-)